+++ components/cascader-panel/useNode.js
@@ -51,7 +51,7 @@ function setNodeElem(node, elem) {
     node.elem = elem;
 }
 
-function useSelectedNodes(config, props, nodes) {
+function useSelectedNodes(config, props, allNodes) {
     const selectedNodes = computed(() => {
         const { emitPath } = config.value;
         const { currentValue, multiple } = props;
@@ -66,7 +66,7 @@ function useSelectedNodes(config, props, nodes) {
                 currentValue,
             );
 
-            const node = getNodeByValue(nodes.value, nodeValue);
+            const node = getNodeByValue(allNodes.value, nodeValue);
 
             if (node) {
                 currentSelectedNodes.push(node);
@@ -79,10 +79,9 @@ function useSelectedNodes(config, props, nodes) {
                 emitPath,
                 currentValue,
             );
-            nodeValues.forEach((nodeValue) => {
-                const node = getNodeByValue(nodes.value, nodeValue);
-                if (node) {
-                    currentSelectedNodes.push(node);
+            allNodes.value.forEach((item) => {
+                if (nodeValues.includes(item.value)) {
+                    currentSelectedNodes.push(item);
                 }
             });
         }
@@ -101,7 +100,7 @@ export default (config, props) => {
         props,
     );
 
-    const { selectedNodes } = useSelectedNodes(config, props, nodes);
+    const { selectedNodes } = useSelectedNodes(config, props, allNodes);
 
     return {
         nodes,