+++ components/select/helper.js
@@ -1,0 +1,2 @@
+let columnIdSeed = 1;
+export const getOptionId = () => columnIdSeed++;