+++ components/tooltip/style/index.js
@@ -1,4 +1,4 @@
 import '../../style/index.less';
-import './index.less';
 import '../../button/style';
 import '../../popper/style';
+import './index.less';