+++ components/tree/useState.js
@@ -7,34 +7,35 @@ export default (props, { emit }) => {
         emit,
         { prop: 'expandedKeys', isEqual: true },
     );
+
     const [currentCheckedKeys, updateCheckedKeys] = useNormalModel(
         props,
         emit,
         { prop: 'checkedKeys', isEqual: true },
     );
+
     const [currentSelectedKeys, updateSelectedKeys] = useNormalModel(
         props,
         emit,
         { prop: 'selectedKeys', isEqual: true },
     );
+
     const { filter, hiddenKeys, filteredExpandedKeys } = useFilter(props);
 
     const hasSelected = (value) => currentSelectedKeys.value.includes(value);
     const hasChecked = (value) => currentCheckedKeys.value.includes(value);
     const hasIndeterminate = (node) => {
-        const value = node[props.valueField];
-        if (hasChecked(value)) {
+        if (hasChecked(node.value)) {
             return false;
         }
         if (node.isLeaf) {
             return false;
         }
         return (
+            props.cascade &&
             Array.isArray(node.children) &&
             node.children.some(
-                (item) =>
-                    hasChecked(item[props.valueField]) ||
-                    hasIndeterminate(item),
+                (item) => hasChecked(node.value) || hasIndeterminate(item),
             )
         );
     };