+++ docs/.vitepress/components/cascader/multiple.vue
@@ -254,7 +254,7 @@
     })
 </script>
 
-<style lang="less" scoped>
+<style scoped>
 .fes-cascader {
     width: 200px;
 }