+++ docs/.vitepress/components/image/preview.vue
@@ -13,4 +13,3 @@ export default {
     },
 };
 </script>
-<style lang="less"></style>