+++ docs/.vitepress/components/select/clearable.vue
@@ -42,7 +42,7 @@ export default {
     }
 };
 </script>
-<style lang="less" scoped>
+<style scoped>
 .fes-select {
     width: 200px;
 }