+++ docs/.vitepress/components/select/disabled.vue
@@ -72,7 +72,7 @@ export default {
     },
 };
 </script>
-<style lang="less" scoped>
+<style scoped>
 .fes-select {
     width: 200px;
 }